+++ ms_python_client/components/events/events_component.py
@@ -1,6 +1,6 @@
 from typing import Any, Mapping, Optional
 
-from ms_python_client.ms_client_interface import MSClientInterface
+from ms_python_client.interfaces.ms_client_interface import MSClientInterface
 
 
 class EventsComponent:
